+++ src/main/java/com/dynatrace/openkit/api/Session.java
@@ -109,10 +109,18 @@ public interface Session extends Closeable {
     WebRequestTracer traceWebRequest(String url);
 
     /**
-     * Reports a BIZ event with a mandatory type and additional attributes
+     * Send a Business Event
+     * 
+     * With sendBizEvent, you can report a business event. These standalone events are being sent detached
+     * from user actions or sessions.
      *
-     * @param type Type of the BIZ event which is mandatory
-     * @param attributes Additional attributes which are passed along side our internal attributes
+     * Note: Business events are only supported on Dynatrace SaaS deployments currently.
+     *
+     * @param type Mandatory event type
+     * @param attributes Must be a valid JSON object and cannot contain functions, undefined,
+     * Infinity, or NaN as values, otherwise they will be removed. Attributes need to be serializable using JSON.stringify.
+     * The resulting event will be populated with the 'attributes'-parameter and enriched with additional properties.
+     * Therefore, even empty objects are valid.
      */
     void sendBizEvent(String type, Map<String, JSONValue> attributes);