+++ src/main/java/com/dynatrace/openkit/core/objects/EventPayloadAttributes.java
@@ -19,14 +19,11 @@ package com.dynatrace.openkit.core.objects;
 public class EventPayloadAttributes {
     public static final String TIMESTAMP = "timestamp";
     public static final String EVENT_KIND = "event.kind";
-    public static final String DT_AGENT_VERSION = "dt.agent.version";
-    public static final String DT_AGENT_TECHNOLOGY_TYPE = "dt.agent.technology_type";
-    public static final String DT_AGENT_FLAVOR = "dt.agent.flavor";
+    public static final String EVENT_PROVIDER = "event.provider";
     public static final String APP_VERSION = "app.version";
     public static final String OS_NAME = "os.name";
     public static final String DEVICE_MANUFACTURER = "device.manufacturer";
     public static final String DEVICE_MODEL_IDENTIFIER = "device.model.identifier";
-    public static final String WINDOW_ORIENTATION = "window.orientation";
 
     public static final String EVENT_KIND_RUM = "RUM_EVENT";
     public static final String EVENT_KIND_BIZ = "BIZ_EVENT";